+++ src/member/member.handler.ts
@@ -13,19 +13,19 @@ export class MemberHandler {
 
   @OnEvent('games.*.created')
   async onGameCreated(game: Game): Promise<void> {
-    await this.memberService.create(game._id, game.owner, {
-      password: undefined,
+    await this.memberService.create(game._id.toString(), game.owner, {
+      password: '',
       ready: false,
     });
   }
 
   @OnEvent('games.*.deleted')
   async onGameDeleted(game: Game): Promise<void> {
-    await this.memberService.deleteGame(game._id);
+    await this.memberService.deleteGame(game._id.toString());
   }
 
   @OnEvent('users.*.deleted')
   async onUserDelete(user: User): Promise<void> {
-    await this.memberService.deleteUser(user._id);
+    await this.memberService.deleteUser(user._id.toString());
   }
 }