+++ src/settlers/player/player.handler.ts
@@ -13,14 +13,12 @@ export class PlayerHandler {
   @OnEvent('games.*.updated')
   async onGameUpdated(game: Game): Promise<void> {
     if (game.started) {
-      await this.playerService.createForGame(game._id);
-    } else {
-      await this.playerService.deleteByGame(game._id);
+      await this.playerService.createForGame(game._id.toString());
     }
   }
 
   @OnEvent('games.*.deleted')
   async onGameDeleted(game: Game): Promise<void> {
-    await this.playerService.deleteByGame(game._id);
+    await this.playerService.deleteByGame(game._id.toString());
   }
 }