+++ src/user/user.schema.ts
@@ -10,7 +10,7 @@ import {
   IsString,
   MaxLength,
 } from 'class-validator';
-import { Document } from 'mongoose';
+import { Document, Types } from 'mongoose';
 import { GLOBAL_SCHEMA_OPTIONS, GlobalSchema, MONGO_ID_ARRAY_FORMAT } from '../util/schema';
 import { IsUrlOrUri } from '../util/url-or-uri.validator';
 
@@ -48,12 +48,12 @@ export class User extends GlobalSchema {
   friends: string[];
 
   @Prop({ transform: () => undefined })
-  passwordHash?: string;
+  passwordHash: string;
 
-  @Prop({ transform: () => undefined })
+  @Prop({ type: String, transform: () => undefined })
   refreshKey?: string | null;
 }
 
-export type UserDocument = User & Document;
+export type UserDocument = User & Document<Types.ObjectId>;
 
 export const UserSchema = SchemaFactory.createForClass(User);